api/requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
pydantic>=2.5.0
python-dateutil>=2.8.2
uuid
json-logging-py>=0.2.0